Add reset ratchet command. (#100)
This diff implements a command to initiate a reset ratchet command. This
works by adding a proxy message that the server will propagate to the
client but remains unenecrypted. This is acceptable because the sigma KX
is designed to operate in an adversarial environment. So even if the
server initiates the command the KX can not be completed unless the
server would have the target's private keys. The reset command is loud
in order to make sure the users see what is going on.

This diff also fixes numerous little issues that were found along these
code paths.