Knock account offline when a duplicate ID comes in
The server doesn't always seem to detect when a client disappears off
the net. This diff adds a mechanism to knock current connections offline
if a duplicate ID comes in. This is annoying when the user may
accidentally have two zkclients open but it beats being locked out of
zkserver for hours at a time.